Feat: Add ads endpoint and pre-version algorithm.

+++ solution/microservices/ad_engine/src/domain/schemas/response/time.rs
@@ -5,5 +5,5 @@
 )]
 pub struct TimeAdvanceResponse {
     #[schema(example = 1)]
-    pub current_date: usize,
+    pub current_date: u32,
 }
